Create a base class Beamline and two child abstract classes BeamlineGoniometer and BeamlineSaxs
